crypto_pipeline/requirements.txt
python-binance==1.0.27
mexc-api==0.0.2
python-okx==0.3.5
psycopg[binary]==3.2.4